changed camera viewing
changed the camera to adjust to its *-1 values on the y when the player is on the anti gravity floor. this way the perspective on both floors will be the same. the spacing from the bottom and top of the screen to the bottom or top of the character (depending on where the character is) will now be consistent.