refactored some code, added game over screen
changed the coroutine for the player jump animation to an async function. added game over screen with play again quit buttons, also shows current score for that run and the start of a high score system being displayed on the side of the screen. currently only displays the current score since there is no save state at this time.