beginning of store page started
store scene is up as a placeholder for now. from main menu hitting play goes into the store which will also act like a player customization page, change skins and spending total score points. after death there is a button that allows the user to go to the store scene